step611

+++ hello_world/6-size.c
@@ -9,8 +9,8 @@ int main(void)
 {
 printf("Size of char: %d\n", sizeof(char));
 printf("\nSize of int: %d\n", sizeof(int));
-printf("\nSize of float: %d\n", sizeof(float));
+printf("\nSize of float: %d\n", sizeof(double));
 printf("\nSize of long int: %d\n", sizeof(long int));
-printf("\nSize of double: %d\n", sizeof(double));
+printf("\nSize of double: %d\n", sizeof(float));
 return (0);
 }
